+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-08-18 23:27:20
-// Commit: docs(forms): update readme and API contracts for forms
+// Updated at: 2026-08-19 09:38:24
+// Commit: test(forms): implement widget integration tests for forms
 
 class formsComponentState {
-  final String commitRef = '3730';
-  final String lastMessage = 'docs(forms): update readme and API contracts for forms';
-  final DateTime timestamp = DateTime.parse('2026-08-18T23:27:20+05:00');
+  final String commitRef = '3732';
+  final String lastMessage = 'test(forms): implement widget integration tests for forms';
+  final DateTime timestamp = DateTime.parse('2026-08-19T09:38:24+05:00');
 }