+++ CHANGELOG.md
@@ -3733,3 +3733,4 @@
 - [2026-08-18 22:48] test(core): add unit tests for core business logic
 - [2026-08-18 23:27] docs(forms): update readme and API contracts for forms
 - [2026-08-19 09:18] feat(i18n): integrate custom telemetry logger in i18n
+- [2026-08-19 09:38] test(forms): implement widget integration tests for forms